requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
markdownify==0.13.1
numpy==1.26.3
pandas==2.2.3
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-jobspy==1.1.80
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.7
tls-client==1.0.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
